requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
Pillow==9.0.0
psycopg==3.0.8
PyJWT==2.3.0
python-dateutil==1.5
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
